Make BeamHasInsertOnConflict associated types injective
